+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-## 0.0.1
+## 0.0.2
 
-* TODO: Describe initial release.
+* Add getConversation method.
+* Add 'sender' field to Message class.